feat: hideEmptyFields is not a component-override prop
commit-id:812cd3fc
