+++ frontend/packages/ux-editor/src/components/config/componentSpecificContent/Summary2/Override/OverrideFields/EmptyTextField.test.tsx
@@ -60,7 +60,6 @@ const onChangeMock = jest.fn();
 const render = () => {
   const override: Summary2OverrideConfig = {
     componentId: component1IdMock,
-    hideEmptyFields: false,
     forceShow: true,
     hidden: false,
   };
